Extract prepareInMainThread to separate entrypoint
To avoid loading fonts and other worker imports in main thread
