+++ pdf/src/index.js
@@ -1,5 +1,5 @@
 import { pdf } from './renderer/index.js'
-import CampPrint, { prepare, prepareInMainThread } from './CampPrint.vue'
+import CampPrint, { prepare } from './CampPrint.vue'
 
 const render = (props = {}) => pdf(CampPrint, props)
-export { render, prepare, prepareInMainThread }
+export { render, prepare }
